Fix/#11: 위도/경도 좌표가 소수가 아닌 정수의 형태로 전달되던 버그 수정

+++ src/main/java/com/dongkyeom/gps/emitter/emitter/application/service/EmitGpsDataService.java
@@ -95,7 +95,7 @@ public class EmitGpsDataService implements EmitGpsDataUseCase {
         if (cell == null) return "";
         return switch (cell.getCellType()) {
             case STRING -> cell.getStringCellValue();
-            case NUMERIC -> String.valueOf((long) cell.getNumericCellValue());
+            case NUMERIC -> String.valueOf(cell.getNumericCellValue());
             case BOOLEAN -> String.valueOf(cell.getBooleanCellValue());
             default -> "";
         };
